refactor: Add getUserPublicProfile endpoint and service method
This commit adds a new endpoint and service method to retrieve the public profile of a user. The getUserPublicProfile endpoint is accessible via a GET request to '/users/:userId'. The service method, getUserPublicProfile, retrieves the user's public profile information and their associated predictions.

The changes include:
- Addition of the getUserPublicProfile method in the UserController class
- Addition of the corresponding route in the user.route.ts file
- Addition of the getUserPublicProfile method in the UserService class

These changes enhance the functionality of the application by allowing users to view the public profiles of other users.